pyproject.toml
[project]
name = "multimodal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.6",
    "dotenv>=0.9.9",
    "google-cloud-aiplatform>=1.128.0",
    "google-genai>=1.52.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-google-vertexai>=3.1.0",
    "langchain-text-splitters>=1.0.0",
    "lxml>=6.0.2",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pip>=25.3",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.0",
    "tiktoken>=0.12.0",
    "unstructured[all-docs]>=0.18.21",
]
